Update case 'b' for printing binary representation of a given number Now able to handle negative numbers as well

+++ main.h
@@ -7,6 +7,7 @@
 #include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 /**
  * struct case_s - selects a case and its function
@@ -48,7 +49,7 @@ int print_hexa_X(unsigned int);
 int print_address(void *);
 int print_unknown(const char *);
 int valid_format(const char *);
-int print_binary(int);
+int print_binary(uintmax_t);
 int base_count(unsigned int, int);
 char *reverse(char *);
 char *rot13(char *);
